requirements.txt
# Enterprise Knowledge Assistant — Backend Requirements
# Install with: pip install -r requirements.txt

# ─── Web Framework ───────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9          # For file uploads

# ─── LangChain Core ──────────────────────────────────────────────────────────
langchain==0.2.6
langchain-core==0.2.10
langchain-community==0.2.6
langchain-google-genai==1.0.7    # Google Gemini integration

# ─── Vector Databases ────────────────────────────────────────────────────────
faiss-cpu==1.8.0                 # FAISS vector store (CPU version)
chromadb==0.5.3                  # Chroma vector store (alternative)

# ─── Document Loaders ────────────────────────────────────────────────────────
pypdf==4.2.0                     # PDF loading
python-docx==1.1.2               # DOCX loading
docx2txt==0.8                    # DOCX text extraction

# ─── Embeddings & Tokenization ───────────────────────────────────────────────
tiktoken==0.7.0                  # Token counting for chunking

# ─── Google AI ───────────────────────────────────────────────────────────────
google-generativeai==0.7.2       # Gemini API client

# ─── Configuration ───────────────────────────────────────────────────────────
python-dotenv==1.0.1             # .env file support
pydantic==2.7.4                  # Data validation
pydantic-settings==2.3.4         # Settings management

# ─── Utilities ───────────────────────────────────────────────────────────────
aiofiles==23.2.1                 # Async file operations
httpx==0.27.0                    # HTTP client
